Make useResults reusable
