+++ src/screens/SearchScreen.js
@@ -1,34 +1,13 @@
 import React from "react";
 import { View, Text, StyleSheet } from "react-native";
 
-import yelp from "../api/yelp";
+import useResults from "../hooks/useResults";
 
 import SearchBar from "../components/SearchBar";
 
 const SearchScreen = () => {
   const [term, setTerm] = React.useState("");
-  const [results, setResults] = React.useState([]);
-  const [errorMessage, setErrorMessage] = React.useState("");
-
-  const searchApi = async (searchTerm) => {
-    try {
-      const response = await yelp.get("/search", {
-        params: {
-          limit: 50,
-          term: searchTerm,
-          location: "berlin",
-        },
-      });
-      setResults(response.data.businesses);
-    } catch (error) {
-      setErrorMessage("Something went wrong");
-    }
-  };
-
-  // Call searchApi when component is first rendered
-  React.useEffect(() => {
-    searchApi("pizza");
-  }, []);
+  const [searchApi, results, errorMessage] = useResults();
 
   return (
     <View>
